requirements.txt
cerebras-cloud-sdk>=1.0.0
chromadb>=0.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
rich>=13.0.0
pydantic>=2.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
